Former-commit-id: 86b685a75d9f94391c40c6404e683efe5d305002 [formerly 86b685a75d9f94391c40c6404e683efe5d305002 [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614]] [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee] [formerly 86b685a75d9f94391c40c6404e683efe5d305002 [formerly 86b685a75d9f94391c40c6404e683efe5d305002 [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614]] [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee] [formerly 75fdd610cc5baf40c0bd05d21c803a1c08844eb1 [formerly 75fdd610cc5baf40c0bd05d21c803a1c08844eb1 [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a]] [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee [formerly f65e8a7253c65b5d59e78ed57e3d73bf01b2e06f]]]]]
Former-commit-id: 75fdd610cc5baf40c0bd05d21c803a1c08844eb1 [formerly 75fdd610cc5baf40c0bd05d21c803a1c08844eb1 [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a]]
Former-commit-id: 33d7681ac1f9c3039bd96917789f2193785d24ef [formerly 4399b9ea393fc7943f119cfc7e52cde0a334b0f9] [formerly 35cc91e58a77947705245c14c2257c9e023c3491 [formerly e280c293be772e3b76431146949849c8b59106f3]] [formerly 4d94e3bedebe8b4a63b18c08a4c412c01d8813b7 [formerly 610366c67eaf484427fc0fef825c2d6e00f2aa45] [formerly 44910ad87f42596202a72a607b73e0ae197d2083 [formerly 44ecd4408b0a495b887d8514b0d31c9c7be47f74]]]
Former-commit-id: c67c6c9b7ff04b0e2ce7fede6eee7f2599f52694 [formerly ff37175b46e444a1df8ac621d5c370e3a6581190] [formerly feaaaa5cd799b046db58064a6e632ae3a0568416 [formerly 6af9e88e9374794fdb017013e2bcc49bbaf728f2]]
Former-commit-id: 6407c7a7a1a0e49308a4d4e364ae4afad6ef5b77 [formerly 9a44f4f5fde9f96ae2d0d1aac7aa3844cc5b3c00]
Former-commit-id: 3b932cef7e014447cc425cfd9adfc92127e69071

+++ edexOsgi/com.raytheon.edex.plugin.obs/src/com/raytheon/edex/plugin/obs/metar/MetarPointDataTransform.java
@@ -62,6 +62,7 @@ import com.raytheon.uf.edex.decodertools.core.DecoderTools;
  * Aug 30, 2013 2298       rjpeter     Make getPluginName abstract
  * Dec 16, 2013 DR 16920   D. Friemdan Fix type of tempFromTenths access.
  * May 14, 2014 2536       bclement    removed TimeTools usage
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -389,8 +390,8 @@ public class MetarPointDataTransform {
 
         SurfaceObsLocation loc = new SurfaceObsLocation(
                 pdv.getString(STATION_NAME));
-        Double lat = pdv.getNumber(LATITUDE).doubleValue();
-        Double lon = pdv.getNumber(LONGITUDE).doubleValue();
+        float lat = pdv.getNumber(LATITUDE).floatValue();
+        float lon = pdv.getNumber(LONGITUDE).floatValue();
         loc.assignLocation(lat, lon);
         loc.setElevation(pdv.getNumber(ELEVATION).intValue());
         mr.setLocation(loc);
